Updated fixed memory model, plus various bugfixes
Rope memory control signals are now decoded much earlier in the cycle to
determine the target address, allowing slower memory devices (such as
the EPCS on the DE0 Nano) enough time to load their data.

The bugfixes are mostly wiring errors discovered by debugging lockups in
when running Colossus 237 -- all were causing the ST2 pulse to not
occur, so errors that *don't* prevent ST2 from happening very well might
be present in the instructions that have been fixed.

The computer now gets through the initialization code and enters an
infinite loop when attempting to invoke the interpeter, due to the lack
of erasable memory.