Added and tied in the Alarms module.
I've placed page 2 of Counter Cell 2 here as page 2 of alarms. It didn't
fit in the counter cells, despite producing a few counter-related
signals -- and it mostly handles monitor stuff, which is a good fit for
what's going on here. As part of this I modernized the timer and scaler
connectors.

There seems to be lots of oscillations in the the simulation. Initial
conditions to correct those to follow.